+++ src/baseclasses/solar_energy/__init__.py
@@ -23,7 +23,8 @@ from .conductivity_measurement import (
 )
 from .eqemeasurement import EQEMeasurement, SolarCellEQECustom
 from .jvmeasurement import JVMeasurement, SolarCellJVCurveCustom
-from .multijunction import ModuleConfiguration, MultijunctionConfiguration, SubcellInfo
+from .module import ModuleConfiguration
+from .multijunction import MultijunctionConfiguration, SubcellInfo
 from .mpp_tracking import MPPTracking, MPPTrackingProperties
 from .mpp_tracking_hysprint_custom import (
     JVData,
